feat: move delete button to bottom & adjust styling

+++ pages/recipe/[recipeId].tsx
@@ -14,7 +14,7 @@ import { useIsLoggedIn } from 'hooks/useIsLoggedIn.hook'
 import RecipeData from '~/components/RecipeData/RecipeData'
 import RecipePlaceholder from '~/components/RecipePlaceholder'
 import FloatingActionButton from '~/components/atoms/FloatingActionButton/FloatingActionButton.component'
-import Button from '~/components/atoms/Button/Button'
+import FloatingWrap from '~/components/atoms/FloatingActionButton/FloatingWrap.component'
 
 type TProps = {
   recipe?: TRecipe
@@ -116,42 +116,44 @@ const RecipePage: NextPage<TProps> = ({ recipe }) => {
       <div className="rooftop flex flex-1 z-10 mt-72 mb-24 pt-8 bg-lightest items-center">
         <div className="flex flex-col flex-1 max-w-5xl pt-0 mx-auto">
           <div className="sticky top-0 pt-8 bg-lightest flex flex-col flex-1 items-center z-10">
-            <div className="flex flex-row">
-              <h2 className="text-darkest text-center px-2">{isFallback ? '...' : recipe.title}</h2>
-              {isLoggedIn && (
-                <Button
-                  className="bg-error absolute right-0 mr-4 lg:mr-0 w-14 h-14 flex items-center justify-center border-4 border-dark"
-                  type="button"
-                  circular
-                  onPress={async () => {
-                    await RecipesApi.delete(recipe.id)
-                    router.back()
-                  }}
-                >
-                  <Image src="/icons/delete.svg" alt="Delete icon" width={24} height={24} />
-                </Button>
-              )}
-            </div>
+            <h2 className="text-darkest text-center px-2">{isFallback ? '...' : recipe.title}</h2>
             <hr className="border-t-4 border-primary w-full" />
           </div>
           {isFallback ? <RecipePlaceholder /> : <RecipeData recipe={recipe} />}
         </div>
       </div>
       {isLoggedIn && (
-        <FloatingActionButton
-          placement="right"
-          onPress={() => {
-            localStorage.setItem('recipe', JSON.stringify(recipe))
-            router.push({
-              pathname: '/add-recipe',
-              query: {
-                prefilled: true
+        <FloatingWrap className="flex flex-row items-center" placement="right">
+          <FloatingActionButton
+            error
+            size="s"
+            className="mr-4"
+            onPress={async () => {
+              // TODO: custom UI
+              // eslint-disable-next-line no-alert, no-restricted-globals
+              const confirmed = confirm(`Zeker dat je ${recipe.title} wilt verwijderen?`)
+              if (confirmed) {
+                localStorage.setItem('recipe', JSON.stringify(recipe))
+                await RecipesApi.delete(recipe.id)
+                router.back()
               }
-            })
-          }}
-        >
-          <Image src="/icons/edit.svg" alt="Edit icon" width={24} height={24} />
-        </FloatingActionButton>
+            }}
+          >
+            <Image src="/icons/delete.svg" alt="Delete icon" width={24} height={24} />
+          </FloatingActionButton>
+          <FloatingActionButton
+            onPress={() => {
+              router.push({
+                pathname: '/add-recipe',
+                query: {
+                  prefilled: true
+                }
+              })
+            }}
+          >
+            <Image src="/icons/edit.svg" alt="Edit icon" width={24} height={24} />
+          </FloatingActionButton>
+        </FloatingWrap>
       )}
     </div>
   )
